+++ NEWS.md
@@ -47,6 +47,7 @@
   there are comments and line breaks under certain circumstances (#905). 
 * rules that add tokens don't break stylerignore sequences anymore (#891).
 * Alignment detection respects stylerignore (#850).
+* `~` causes now indention, like `+`, `-`, `|` etc. (#902).
 * `Warning: Unknown or uninitialised column:` was fixed (#885).
 * Unaligned expressions with quoted key (e.g. `c("x" = 2)`) are now correctly
   detected (#881).